% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_extract_cs_embedding.r
\name{run_Structformer_extract_cs_embedding}
\alias{run_Structformer_extract_cs_embedding}
\title{Calculate the embedding for spots/cells}
\usage{
run_Structformer_extract_cs_embedding(
  seu_obj,
  save_checkpoint_path,
  pretrained_model,
  envir_path,
  with_gpu = FALSE
)
}
\arguments{
\item{seu_obj}{The Seurat object contains the data for spots/cells.}

\item{save_checkpoint_path}{The save path for the Structformer trained model.}

\item{pretrained_model}{The default is Structformer_BERT. Options include Structformer_BERT or Geneformer. Use this function when the save_checkpoint_path parameter is NULL.}

\item{envir_path}{The python env path.}

\item{with_gpu}{Default: FALSE. Specify TRUE or FALSE to indicate whether to use a GPU.}

\item{pretrained_model_path}{The save path for the pre-trained BERT model. Use this function when the save_checkpoint_path parameter is NULL.}
}
\value{
The seurat object
}
\description{
Run the Structformer trained model or a pre-trained BERT model to extract the embeddings of the spots/cells.
}
