% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_extract_gene_embedding.r
\name{run_Structformer_extract_gene_embedding}
\alias{run_Structformer_extract_gene_embedding}
\title{Extract gene embedding}
\usage{
run_Structformer_extract_gene_embedding(
  gene_list,
  save_checkpoint_path,
  pretrained_model_path = NULL,
  res = 0.7,
  envir_path,
  with_gpu = FALSE
)
}
\arguments{
\item{gene_list}{The gene list input.}

\item{save_checkpoint_path}{The save path of Structformer trained model.}

\item{pretrained_model_path, }{The pre-trained model saved path.}

\item{res}{The resoultion to distinguish clusters.}

\item{envir_path}{The python env path.}

\item{with_gpu}{Defulat FALSE, TRUE/FALSE, use GPU or not.}
}
\value{
The relative distance of predicted single cells or spots with TLS prototype and non-TLS prototype, the prediction label of whether a single cell or spot belong to TLS region.
}
\description{
Run Structformer trained model to extract key genes of a group spots/cells.
}
