% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_extract_gene_module.r
\name{run_Structformer_extract_gene_module}
\alias{run_Structformer_extract_gene_module}
\title{Extract gene modules}
\usage{
run_Structformer_extract_gene_module(
  seu_obj,
  save_checkpoint_path,
  envir_path,
  target_name,
  phenotype_num,
  hvg_num = 3000,
  res,
  genes_representor,
  with_gpu = FALSE
)
}
\arguments{
\item{seu_obj}{The seurat object which will be used for predicting.}

\item{save_checkpoint_path}{The save path of Structformer trained model.}

\item{envir_path}{The python env path.}

\item{target_name}{The colname of which you want to predict}

\item{phenotype_num}{The number represents the phenotype in the target_name column}

\item{hvg_num}{The number of top gene you want to get}

\item{res}{The resoultion to distinguish clusters.}

\item{genes_representor}{The file path of genes were used in the pre-training the gene word encoder.}

\item{with_gpu}{Defulat FALSE, TRUE/FALSE, use GPU or not.}
}
\value{
The relative distance of predicted single cells or spots with TLS prototype and non-TLS prototype, the prediction label of whether a single cell or spot belong to TLS region.
}
\description{
Run Structformer trained model to extract key genes of a group spots/cells.
}
