% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_extract_keygenes_fulldoc.r
\name{run_Structformer_extract_keygenes}
\alias{run_Structformer_extract_keygenes}
\title{Extract key genes of a group of spots/cells}
\usage{
run_Structformer_extract_keygenes(
  seu_obj,
  save_checkpoint_path,
  envir_path,
  target_name,
  phenotype_num,
  metadata = TRUE,
  top_n = 60,
  with_gpu = FALSE
)
}
\arguments{
\item{seu_obj}{The seurat object which will be used for predicting.}

\item{save_checkpoint_path}{The save path of Structformer trained model.}

\item{envir_path}{The python env path.}

\item{target_name}{The colname of which you want to predict.}

\item{phenotype_num}{The number represents the phenotype in the target_name column.}

\item{metadata}{Whether use metadata of Seurat object}

\item{top_n}{The number of top gene you want to get.}

\item{with_gpu}{Defulat FALSE, TRUE/FALSE, use GPU or not.}
}
\value{
Key genes of a group spots/cells
}
\description{
Run Structformer trained model to extract key genes of a group spots/cells.
}
