% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gene_network.R
\name{run_Structformer_gene_network}
\alias{run_Structformer_gene_network}
\title{Gene correlation analysis and network construction}
\usage{
run_Structformer_gene_network(
  gene_list,
  save_checkpoint_path,
  res = 0.7,
  link_thres = 0.6,
  col_list = NULL,
  out_dist = 2.5,
  envir_path = NULL,
  with_gpu = FALSE
)
}
\arguments{
\item{gene_list}{The seurat object which will be used for predicting.}

\item{save_checkpoint_path}{The save path of Structformer trained model.}

\item{res}{The resoultion to distinguish clusters.}

\item{link_thres}{The threshold of cosine similarity score of two genes' linking.}

\item{col_list}{The color list for gene modules.}

\item{out_dist}{The distance of label between circle.}

\item{envir_path}{The python env path.}

\item{with_gpu}{Defulat FALSE, TRUE/FALSE, use GPU or not.}
}
\value{
The relative distance of predicted single cells or spots with TLS prototype and non-TLS prototype, the prediction label of whether a single cell or spot belong to TLS region.
}
\description{
Run Structformer trained model to extract key genes of a group spots/cells.
}
