% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pred_Structformer.r
\name{run_Structformer_pred}
\alias{run_Structformer_pred}
\title{Predict single cell or spots whether belong to TLS region}
\usage{
run_Structformer_pred(
  seu_obj,
  pretrained_model_path,
  save_checkpoint_path,
  envir_path,
  pretrained_model = "Structformer_BERT",
  sen_len = 260,
  data_type = "sc_st",
  metadata = FALSE,
  class_num = 2,
  with_gpu = TRUE
)
}
\arguments{
\item{seu_obj}{The seurat object which will be used for predicting.}

\item{pretrained_model_path}{The pre-trained gene word encoder model saved path.}

\item{save_checkpoint_path}{The save path of Structformer trained model.}

\item{envir_path}{The python env path.}

\item{pretrained_model}{Default is TLSPredictor_BERT. Structformer_BERT or geneformer.}

\item{sen_len}{Default is 260. The sentence length, the generated sentences length will be minus or equal this parameter. If the gene expression level is zero, the gene will not be invovled.}

\item{data_type}{Default is sc_st, the other is bulk.}

\item{metadata}{Use metadata of Seurat object, default is FALSE.}

\item{class_num}{The classification number, default is 2, it can support multiple classification.}

\item{with_gpu, }{TRUE/FALSE, use GPU or not.}
}
\value{
The relative distance of predicted single cells or spots with TLS prototype and non-TLS prototype, the prediction label of whether a single cell or spot belong to TLS region.
}
\description{
Run Structformer trained model to calculate relative distance of single cells or spot with reference TLSs or Non-TLSs spots and infer whether ir belong to TLS region.
}
