% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_training_Structformer.r
\name{run_Structformer_train}
\alias{run_Structformer_train}
\title{Train Structformer on spatial transcriptomics data}
\usage{
run_Structformer_train(
  seu_obj,
  pretrained_model = "Structformer_BERT",
  sen_len = 260,
  pretrained_model_path,
  save_checkpoint_path,
  batch_size,
  train_K,
  train_Q,
  train_episodes,
  val_episodes = 30,
  val_steps = 10,
  metadata = FALSE,
  reproduce = FALSE,
  set_seed = 42,
  target_name,
  envir_path,
  with_gpu = TRUE
)
}
\arguments{
\item{seu_obj, }{The seurat object which will be used for training.}

\item{pretrained_model, }{Default is Structformer_BERT. Structformer_BERT or geneformer.}

\item{sen_len, }{Default is 260. The sentence length, the generated sentences length will be minus or equal this parameter. If the gene expression level is zero, the gene will not be invovled.}

\item{pretrained_model_path, }{The pre-trained model saved path.}

\item{save_checkpoint_path, }{The save path of model.}

\item{batch_size, }{Batch size of training process.}

\item{train_K, }{Support set numbers.}

\item{train_Q, }{Query set numbers.}

\item{train_episodes, }{Training episodes.}

\item{val_episodes, }{Validation is performed after the set number of episodes. The default is 30.}

\item{val_steps, }{After val_steps, the accuracy of the current model is compared to the previous model. If the accuracy score improves, the current model will be saved..}

\item{metadata, }{Use metadata of Seurat object.}

\item{reproduce, }{Whether reproduce training process.}

\item{set_seed, }{Set a seed.}

\item{target_name, }{The colname of which you want to predict.}

\item{envir_path, }{The python env path.}

\item{with_gpu, }{TRUE/FALSE, use GPU or not.}
}
\value{
The trained Structformer model and support vector.
}
\description{
Train the Structformer that is based on metrics learing few-shot learning to deal with the unblance problem.
Beacuse The number of TLSs spots for training is generally much smaller that non-TLSs spots.
}
